+++ client/.eslintrc
@@ -14,5 +14,8 @@
         ]
       }
     }
+  },
+  "rules": {
+    "@typescript-eslint/ban-ts-comment": "off"
   }
 }
