+++ client/src/components/navigation/TopNav/components/index.ts
@@ -1,0 +1,1 @@
+export {default as HeaderLinks} from './links.top-nav';
